added faq route, theme color change
