+++ src/components/expandable/expandable.module.css
@@ -7,7 +7,7 @@
     flex-direction: row;
     justify-content: space-between;
     align-items: center;
-    background-color: whitesmoke;
+    background-color: white;
     color: black;
     height: 3rem;
     padding: 0 1.5rem 0 1.5rem;
@@ -39,7 +39,7 @@
         flex-direction: row;
         justify-content: space-between;
         align-items: center;
-        background-color: whitesmoke;
+        background-color: white;
         color: black;
         height: 4rem;
         border-radius: 0.5rem;
@@ -71,7 +71,7 @@
         flex-direction: row;
         justify-content: space-between;
         align-items: center;
-        background-color: whitesmoke;
+        background-color: white;
         color: black;
         height: 6rem;
         border-radius: 0.5rem;
@@ -107,7 +107,7 @@
     to {
         opacity: 1;
         transform: translateY(-10px);
-        background-color: white;
+        background-color: aliceblue;
 
     }
 }
