+++ src/components/nav/navbar.module.css
@@ -1,0 +1,19 @@
+.container {
+    display: flex;
+    justify-content: flex-end;
+    align-items: center;
+    width: 100%;
+    height: 3rem;
+    background-color: powderblue;
+}
+
+.link {
+    color: black;
+    padding: 1rem;
+    text-decoration: none;
+}
+
+.link:hover {
+    text-decoration: underline;
+    color: navy
+}
